+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+* Event self-announce feature at configurable time before event starts
 * Blueprint (Preview via WordPress Playground)
 
 ## [0.3.5] - 2025-01-03